Extract random address generation into a trait
